requirements.txt
openai-chatkit
fastapi
uvicorn[standard]
httpx
python-dotenv
PyPDF2
python-docx
openpyxl
